pyproject.toml
[project]
name = "deco"
version = "0.1.0"
description = ""
authors = [
    {name = "dssera",email = "adesser001@mail.ru"}
]
readme = "README.md"
requires-python = "3.12.9"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "dishka (>=1.5.3,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false